Move gcc.py to lib/
