+++ pwndbg/gdblib/typeinfo.py
@@ -10,8 +10,8 @@ import sys
 
 import gdb
 
-import pwndbg.gcc
 import pwndbg.gdblib.events
+import pwndbg.lib.gcc
 import pwndbg.lib.memoize
 import pwndbg.lib.tempfile
 
@@ -205,7 +205,7 @@ def compile(filename=None, address=0):
     objectname = os.path.splitext(filename)[0] + ".o"
 
     if not os.path.exists(objectname):
-        gcc = pwndbg.lib.gcc.which()
+        gcc = pwndbg.lib.gcc.which(pwndbg.gdb.arch)
         gcc += ["-w", "-c", "-g", filename, "-o", objectname]
         try:
             subprocess.check_output(gcc)
